server/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
pycryptodome==3.11.0
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
schedule==1.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
